+++ db_utils/database.py
@@ -1,0 +1,5 @@
+from pymongo import MongoClient
+
+
+client = MongoClient(port=27018)
+db = client.todolist
